Support for formatting and header spell-check
Added support for formatting - in other words, bold text will be
highlighted in bold, italic text in italic and so on.

Enhanced spell-checking so that headings can be spell-checked. It doesn't
make much sense to me to spell-check paragraphs, but not headings. Text
within B<> and I<> can now also be spell-checked too. Again, it seems
strange not to spell-check text simply because it is bold or italic.

These features are controlled by separate variables: perl_pod_formatting
and perl_pod_spellcheck_headings which would normally be set from your
vimrc file. The features are disabled by default to maintain compatibility.