highlight TODO: in addition to TODO
perl-support.vim plugin adds ':' to 'iskeyword' option.
At least, its older releases added it (can't find it in the latest
version). I like it because it allows me to move around using */# commands
quickly.
But if this option is configured this way, syntax/perl.vim doesn't
highlight 'TODO:' in comments, because TODO: is considered a complete
keyword.

Maybe I should've replaced 'syn keyword perlTodo' with 'syn match', but
it's not too much of copy-pasting (see the patch), and that's what
syntax/help.vim does for NOTE: lines in vim help files.