time-it.vim: Profile an entire file
This script was only profiling the first screenful of text from the
source file. Now it scrolls through and highlights the entire file.

The profiling report still depends on the (terminal) window size.
A larger window size results in fewer page scrolls and less work for
Vim (fewer matches in the report).