Catch all exceptions when system() fails
Until now it was only catching when Vim was run in restricted mode. system()
can fail in other situations that we can't do much about, like when noexec
is in effect for the user. Since this is not critical functionality, we
should just go for graceful degradation here and just skip it.