Beat podFormat into submission
...kinda.

This is a revision of issue #237.  It properly constrains podFormat
regions to stay within their end pattern, largely by flipping the order
of the podFormat region definition lines *le sigh*.  We also drop the
irrelevant `me=...` offset, as `me` offsets are not
honored for start patterns.

We also define a new style, podFormatError, and use it to handle the
`I<>`/etc cases, making it unnecessary for us to muck around with the
start/end patterns.

Things not addressed: applying podFormatError to text inside podFormat
(e.g. `L<...>` done incorrectly) or to `I<< >>`-style sequences,
considering if we really want `@NoSpell`, etc.

This change causes the following POD to render correctly, at least for
me.  It seemed like a good place to start; I'll address any case not
caught if pointed out or discovered.

    If files are not specified for searching, either on the command
    line or piped in with the C<-x> option, I<ack> delves into
    subdirectories selecting files for searching.

    I<ack> is intelligent about the files it searches.  It knows about
    certain file types, based on both the extension on the file and,
    in some cases, the contents of the file.  These selections can be
    made with the B<--type> option.

    With no file selection, I<ack> searches through regular files that
    are not explicitly excluded by B<--ignore-dir> and B<--ignore-file>
    options, either present in F<ackrc> files or on the command line.
    ```

    B<I<Please>>?  And I<B<Please>> la la la
    B< I<Please>>?  And I<B<Please>> la la la
    B< I<Please>>?  And I<B<Please>> la la la
    B<< I<Please> >>?  And I<< B<< Please >> >> la la la

    beep E<acute> boop Z<> bap

    The I<> and B<> properly I<terminate> their highlighting, but the F<> and C<> do not.  At least on my setup.

    This should be podFormat then podFormatError: I<I<>> and I<< I<> >>, right?