Properly parallelize corpus tests
Before, we were using test methods directly in our child worker
processes.  I thought that the Test framework had a way to communicate
between a family of processes (I don't know what led me to believe that,
though), but I've observed inconsistencies when trying to issue our
final "corpus failed" help message.  It's easier and arguably more
correct just to do the Vim coloration and folding in the children, and
then send the values back up to the parent to apply the tests.