Fix package folding with one-line subs
There was a bug with folding braced packages where two or more subroutines
defined on a single line would have their closing braces treated as
the closing brace for the package block (the details of why two or more
and not one is up to a trick we use called perlBraces in
syntax/perl.vim, which is how we handle nested blocks).  This didn't
happen if you had anonymous subroutine folding turned on, as the sub
folding rule there was a little more flexible.  Properly fixing
the sub rule is difficult and has a lot of potential to break existing
setups, so until we figure that out, here's a fix for braced packages.

The fix just assumes that the ending brace for a package will stand
on its own line and have the same amount of indentation as the package
keyword that introduced it - this mirrors expectations of Perl code
in the real world