Fix folding of DATA sections with a empty final line
The end of a folded syntax region is excluded from the fold if it is an
empty line.  For DATA sections this is an empty final line in the file.

This fix works around the problem by exploiting an unterminated syntax
region, with an end pattern that won't be found, to run to the end of
the file and include this empty line.