Improve heredoc highlighting (#293)
* Highlight system command heredocs

* Support backslash escaped heredoc delimiters

print <<\EOS;
single quoted text
EOS

* Remove useless z\(\) capture groups in heredoc patterns

* Merge perlHeredocStart and perlIndentedHeredocStart syntax groups

The distinction is not important because only the appropriate form can
be contained by the perl*Heredoc regions.

* Remove perlAutoload syntax group

This is such an old an infrequently used idiom that it's not worth
supporting.

It's also so catastrophically buggy, nuking the rest of the file's
highlighting, that I don't believe anyone's utilising it.  I'd really
rather not bother fixing it.